+++ packages/core/ppx/observe_ppx.ml
@@ -1,4 +1,5 @@
 let () =
   Deriver.register ();
   Ppxlib.Reserved_namespaces.reserve "observe";
-  Value_extension.register ()
+  Ppxlib.Driver.register_transformation "observe"
+    ~rules:(Value_extension.rules @ Log_extension.rules)