+++ test/core/ppx/logging_emit_shape.ml
@@ -1,0 +1,1 @@
+let _ = [%observe.emit Observe.Level.Info]