+++ test/core/ppx/logging_unknown_form.ml
@@ -1,0 +1,1 @@
+let _ = [%observe.info value 1]